requirements.txt
Django==5.0.1
supabase==2.9.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==10.4.0
requests==2.31.0
websockets>=13.0
